# Conflicts:
#	SConstruct
#	extern/fftw/fftw-3.3.8.tar.bz2
#	extern/libapng/libpng-1.6.29-apng.patch.tar.bz2
#	extern/libapng/libpng-1.6.29.tar.bz2
#	extern/libevdev/libevdev-1.5.6.tar.bz2
#	extern/libffi/libffi-3.3-rc0.tar.bz2
#	extern/libhidapi/hidapi-hidapi-0.8.0-rc1.tar.bz2
#	extern/libjpeg/libjpeg-turbo-2.0.3.tar.bz2
#	extern/libltdl/libtool-2.4.6.tar.bz2
#	extern/libogg/libogg-1.3.2.tar.bz2
#	extern/libopenhmd/OpenHMD-master.tar.bz2
#	extern/libpng/libpng-1.6.29.tar.bz2
#	extern/libsigsegv/libsigsegv-2.11.tar.bz2
#	extern/libtheora/libtheora-1.1.1.tar.bz2
#	extern/libusb/libusb-1.0.22.tar.bz2
#	extern/libvorbis/libvorbis-1.3.5.tar.bz2
#	extern/shared/gnuconfig.tar.bz2
#	extern/smalltalk/smalltalk-3.2.5.tar.bz2
#	extern/sndio/sndio-1.6.0.tar.bz2
#	extern/soundtouch/soundtouch-2.1.1.tar.bz2
#	extern/zlib/zlib-1.2.11.tar.bz2

+++ src/launcher/gui/src/gui/engine/deglPanelEngine.cpp
@@ -259,7 +259,7 @@ long deglPanelEngine::onListModulesDblClick( FXObject *sender, FXSelector select
 }
 
 long deglPanelEngine::onListModulesHeaderClicked( FXObject*, FXSelector, void *pdata ){
-	int colon = ( intptr_t )pdata;
+	int colon = ( int )( intptr_t )pdata;
 	
 	if( colon == 0 ){
 		if( pSortListModules == elmsNameAsc ){
